\name{VIMGUI}
\alias{VIMGUI}
\title{GUI for Visualization and Imputation of Missing Values}
\description{
    Graphical user interface for visualization and imputation of 
    missing values.
}
\usage{
VIMGUI(startupObject=NULL)
}

\arguments{
    \item{startupObject}{Object loaded at the start of the GUI}
}
\details{
  Details about handling survey objects follow soon.    
}
\author{Daniel Schopfhauser}
\keyword{multivariate}
\keyword{hplot}

