% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIMGUI.R
\name{VIMGUI}
\alias{VIMGUI}
\title{GUI for Visualization and Imputation of Missing Values}
\usage{
VIMGUI(startupObject = NULL)
}
\arguments{
\item{startupObject}{Object loaded at the start of the GUI}
}
\description{
Graphical user interface for visualization and imputation of missing values.
}
\details{
Details about handling survey objects follow soon.
}
\author{
Daniel Schopfhauser
}
\references{
M. Templ, A. Alfons, P. Filzmoser (2012) Exploring incomplete
data using visualization tools.  \emph{Journal of Advances in Data Analysis
and Classification}, Online first. DOI: 10.1007/s11634-011-0102-y.

A. Kowarik, M. Templ (2016) Imputation with
R package VIM.  \emph{Journal of
Statistical Software}, 74(7), 1-16
}
\keyword{hplot}
\keyword{multivariate}

