\name{VIMGUI}
\alias{VIMGUI}
\title{GUI for Visualization and Imputation of Missing Values}
\usage{
  VIMGUI(startupObject = NULL)
}
\arguments{
  \item{startupObject}{Object loaded at the start of the
  GUI}
}
\description{
  Graphical user interface for visualization and imputation
  of missing values.
}
\details{
  Details about handling survey objects follow soon.
}
\author{
  Daniel Schopfhauser
}
\keyword{hplot}
\keyword{multivariate}

